carry name with parameters when traversing spec tree
previously, as the spec tree was navigated any given function was
ignorant of the name of the spec's parent variable; it knew only the
children.  While this approach obviated passing extra name information
along, it made it impossible to pass a fully qualified name to any
callbacks.

the full name is now passed along as a Name object containing an
ordered list of the components in a variable name as well as the index
of the last valid component. The name object is cloned at each level;
the new object gets an independent copy of the index, but shares the
same list of names as its parent.  This is more efficient than copying
the list, and because the parent's index is untouched removes the
need to independently increment and decrement the index as levels
in the tree are traversed.