use data element name object (rather than spec table) to uniquely specify state
if a specification table is reused (for example during specification validation
the internal validate_spec table is used multiple times) then state information
was being mixed, as the specification table was used as the unique key into
the state table.  Use instead the data element name object, which will be unique.