nil positional parameters are now by default not allowed.
The default behavior was to allow nil positional arguments.  This is
inconsistent with the behavior for table arguments (e.g., optional
defaults to false), and caused confusion.

The new default behavior is that positional arguments may not be nil.
A new specification, allow_nil, may be used to indicate that nil
values are permitted.  The not_nil specification is now redundant and
deprecated.